here you go nigga

+++ includes/roblox/Humanoid.h
@@ -5,7 +5,7 @@ namespace RBX {
 	public : 
 	    Humanoid() = default;
 	    explicit Humanoid(uintptr_t instance_address) : Instance(instance_address){};
-	    void set_walkspeed(int walkspeed) const;
+	    void set_walkspeed(float walkspeed) const;
 	private :
 	    bool class_valid() const;
     };
